Move most methods out from ServerAdapter class
