+++ src/main/java/Server/Network/ServerState.java
@@ -1,0 +1,10 @@
+package Server.Network;
+
+public abstract class ServerState {
+
+    public ServerState() {
+
+    }
+
+
+}
